+++ next.config.js
@@ -20,6 +20,7 @@ const securityHeaders = [
 ];
 
 const nextConfig = {
+  output: 'standalone',
   async headers() {
     return [
       {